now float can be use in length cut

+++ src/com/ipinyou/fmpegWrapper/operator/fmpegOperators/SetEndPointOperator.java
@@ -6,7 +6,7 @@ import com.ipinyou.fmpegWrapper.operator.Selector;
 
 public class SetEndPointOperator extends FfmpegOperator {
 
-	public SetEndPointOperator(Selector selector, int endTime) {
+	public SetEndPointOperator(Selector selector, double endTime) {
 		super("t", selector, "" + endTime, SegmentType.INPUT, SegmentType.OUTPUT);
 	}
